backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.10.0
python-dotenv==1.0.0
pyjwt[crypto]>=2.10.1
bcrypt==4.1.2
gunicorn==21.2.0
httpx[http2]>=0.26,<0.29
websockets>=13.0